refactor: Replace home page with timer
